+++ test/test_xmlmap/test_ead.py
@@ -26,8 +26,6 @@ from eulxml.utils.compat import u
 from eulxml.xmlmap import load_xmlobject_from_file, load_xmlobject_from_string
 from eulxml.xmlmap import eadmap
 
-proxy_required = skipIf('HTTP_PROXY' not in os.environ,
-                        'Schema validation test requires an HTTP_PROXY')
 
 
 class TestEad(unittest.TestCase):
@@ -55,7 +53,7 @@ class TestEad(unittest.TestCase):
         self.assert_("12 oversized papers (OP)" in self.ead.physical_desc)
         self.assert_("materials relating to Irish poet Seamus Heaney" in u(self.ead.abstract))
 
-    @proxy_required
+    
     def test_validation(self):
         # EAD objects can now be validated aginst XSD schema
         self.assertTrue(self.ead.schema_valid())